Changes to be committed: 	new file:   .gitignore 	new file:   Makefile 	modified:   ai.cpp 	new file:   ai.o 	new file:   ait.py 	new file:   ait.pyc 	deleted:    aitest 	new file:   feedforward_benchmark 	new file:   hans.o 	modified:   hanslayer.cpp 	new file:   hanslayer.o 	modified:   hanslayer_dense.cpp 	new file:   hanslayer_dense.o 	deleted:    pyaitest 	modified:   pyaitest.py 	modified:   pyaitest.pyc
state of code is good for a benchmark.
feedforward can be compared:
cpp version hANS is implemented to go over 5000 iterations of a feedforward of a randomly setted up network at initialization
python version has no object-oriented structure but uses numpy to perform similar operations on a fixed set of weights but these were also once set by the cpp-version hANS, so calculations are actually not completely equal but comparable in that sense that the weights are both times a uniformly distrubed set!

also look for the benchmark file -> it shows one run of each using my host-computer with an amdfx6300 (six-core, 3,5GHz)